pyproject.toml
[tool.poetry]
name = "sa-project"
version = "0.1.0"
description = ""
authors = ["Korionov Dmitriy <dmytro.korniinov@wastehero.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
uvicorn = "^0.30.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.4.0"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
faker = "^26.0.0"
factory-boy = "^3.3.0"
ruff = "^0.5.1"
django-split-settings = "^1.3.2"
django-environ = "^0.11.2"
gunicorn = "^22.0.0"
pre-commit = "^3.7.1"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
